requirements.txt
fitz  # PyMuPDF
numpy
faiss-cpu  # For CPU-based FAISS
transformers
torch
regex
